adminPage/requirements.txt
fastapi
pydantic
uvicorn
prometheus-fastapi-instrumentator
python-dotenv
langchain_community
langchain
pymupdf
boto3